refactor: extract strategy classes from engine, fix_pipeline, and result_handler (fixes #518)
Extract collaborator classes from the three largest files:
- BlockDecision/evaluate_review_blocking → engine/blocking.py
- SerialDispatcher/ParallelDispatcher → engine/dispatch.py
- CoderReviewerLoop → nightshift/coder_reviewer.py
